third_party/WebKit/LayoutTests/platform/win/http/tests/security/suborigins/crossorigin/suborigin-cors-font-face-expected.txt

CONSOLE ERROR: Access to Font at 'http://127.0.0.1:8000/security/resources/cors-font.php?' from origin 'http-so://foobar.127.0.0.1:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Suborigin' header has a value '' that is not equal to the supplied suborigin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Font at 'http://127.0.0.1:8000/security/resources/cors-font.php?&cors=false' from origin 'http-so://foobar.127.0.0.1:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Suborigin' header has a value '' that is not equal to the supplied suborigin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS FontFace: ACAO: http://127.0.0.1:8000 
PASS FontFace: ACAO: * 
PASS FontFace: CORS-ineligible resource 
Harness: the test ran to completion.
